requirements.txt
flask==2.3.3
pandas==2.0.3
flask-cors==4.0.0
python-dotenv==1.0.0